# Transactions endpoint `/transactions` Performance Improvement