--- a/app/views/transactions/_transaction.html.erb
+++ b/app/views/transactions/_transaction.html.erb
@@ -2,12 +2,13 @@
 
 <% transaction = entry.entryable %>
 
-<%= turbo_frame_tag dom_id(entry) do %>
-  <%= turbo_frame_tag dom_id(transaction) do %>
-    <div class="grid grid-cols-12 items-center text-primary text-sm font-medium p-4 lg:p-4
-                <%= @focused_record == entry || @focused_record == transaction ?
-                    "border border-gray-900 rounded-lg" : "" %>
-                <%= entry.excluded ? "opacity-50 text-gray-400" : "" %>">
+<% cache [entry, entry.account, view_ctx, @focused_record, entry.account.family.entries_cache_version] do %>
+  <%= turbo_frame_tag dom_id(entry) do %>
+    <%= turbo_frame_tag dom_id(transaction) do %>
+      <div class="grid grid-cols-12 items-center text-primary text-sm font-medium p-4 lg:p-4
+                  <%= @focused_record == entry || @focused_record == transaction ?
+                      "border border-gray-900 rounded-lg" : "" %>
+                  <%= entry.excluded ? "opacity-50 text-gray-400" : "" %>">
 
       <div class="pr-4 lg:pr-10 flex items-center gap-3 lg:gap-4 col-span-8 <%= view_ctx == "global" ? "lg:col-span-8" : "lg:col-span-6" %>">
         <%= check_box_tag dom_id(entry, "selection"),
@@ -112,6 +113,7 @@
           <% end %>
         </div>
       <% end %>
-    </div>
+      </div>
+    <% end %>
   <% end %>
 <% end %>