@adeyinkaezra123 2025-07-13
> Note about setup: Benchmarking results may vary based on the data in the DB. The best way to benchmark is setting up a production-mirror with anonymized / scrubbed data, which is a lengthy process that we cannot provide in an OSS context. The best alternative we have is synthetic demo data, which can be generated via rake demo_data:default. Benchmarks can still be run on this small demo family, but time/iteration will be much smaller and performance issues caused by inefficient queries running on large datasets will not show up in the results.

I followed the instructions from the benchmarking setup [PR](https://github.com/maybe-finance/maybe/pull/2366
), which I believe uses the `derailed_benchmark` gem. I had written a simple custom benchmark myself, but once I saw the existing rake task, I went with that instead.

I ran the `demo_data:default` rake task to populate the maybe_production database, and here’s a snapshot of the resulting table sizes:


```
              table_name               | estimated_rows 
---------------------------------------+----------------
 public.balances                       |          28627
 public.entries                        |           8018
 public.transactions                   |           7487
 public.sessions                       |           6176
 public.holdings                       |           4406
 public.trades                         |            526
 public.transfers                      |            433
 public.schema_migrations              |            195
 public.categories                     |             21
 public.syncs                          |             20
 public.accounts                       |             20
 public.budget_categories              |             16
 public.valuations                     |              5
 public.investments                    |              5
 public.depositories                   |              4
 public.securities                     |              3
 public.loans                          |              3
 public.users                          |              2
 public.credit_cards                   |              2
 public.vehicles                       |              2
...
(other tables are with only 0-5 rows)
```

3. I ran into multiple N+1 query issues just by hitting the `/transactions` endpoint. I was using the Prosopite gem, which flagged several repeated queries. Here’s an excerpt from the logs: from the main branch

```log
12:16:09 web.1    | N+1 queries detected:
12:16:09 web.1    |   SELECT "transactions".* FROM "transactions" WHERE "transactions"."id" = $1 LIMIT $2
12:16:09 web.1    |   SELECT "transactions".* FROM "transactions" WHERE "transactions"."id" = $1 LIMIT $2
12:16:09 web.1    |   SELECT "transactions".* FROM "transactions" WHERE "transactions"."id" = $1 LIMIT $2
12:16:09 web.1    | Call stack:
12:16:09 web.1    |   app/models/transfer.rb:57:in 'Transfer#to_account'
12:16:09 web.1    |   app/models/transfer.rb:100:in 'Transfer#categorizable?'
12:16:09 web.1    |   app/views/transactions/_transaction_category.html.erb:4
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:96
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:6
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:5
12:16:09 web.1    |   app/views/entries/_entry.html.erb:3
12:16:09 web.1    |   app/views/transactions/index.html.erb:84
12:16:09 web.1    |   app/helpers/entries_helper.rb:23:in 'block (2 levels) in EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/helpers/entries_helper.rb:22:in 'block in EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Array#reverse_each'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Enumerator#each'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Enumerable#map'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/views/transactions/index.html.erb:83
12:16:09 web.1    |   app/controllers/application_controller.rb:17:in 'ApplicationController#n_plus_one_detection'
12:16:09 web.1    |   app/controllers/concerns/localize.rb:17:in 'Localize#switch_timezone'
12:16:09 web.1    |   app/controllers/concerns/localize.rb:12:in 'Localize#switch_locale'
12:16:09 web.1    | 
12:16:09 web.1    | N+1 queries detected:
12:16:09 web.1    |   SELECT "entries".* FROM "entries" WHERE "entries"."entryable_id" = $1 AND "entries"."entryable_type" = $2 LIMIT $3
12:16:09 web.1    |   SELECT "entries".* FROM "entries" WHERE "entries"."entryable_id" = $1 AND "entries"."entryable_type" = $2 LIMIT $3
12:16:09 web.1    |   SELECT "entries".* FROM "entries" WHERE "entries"."entryable_id" = $1 AND "entries"."entryable_type" = $2 LIMIT $3
12:16:09 web.1    | Call stack:
12:16:09 web.1    |   app/models/transfer.rb:57:in 'Transfer#to_account'
12:16:09 web.1    |   app/models/transfer.rb:100:in 'Transfer#categorizable?'
12:16:09 web.1    |   app/views/transactions/_transaction_category.html.erb:4
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:96
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:6
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:5
12:16:09 web.1    |   app/views/entries/_entry.html.erb:3
12:16:09 web.1    |   app/views/transactions/index.html.erb:84
12:16:09 web.1    |   app/helpers/entries_helper.rb:23:in 'block (2 levels) in EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/helpers/entries_helper.rb:22:in 'block in EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Array#reverse_each'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Enumerator#each'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'Enumerable#map'
12:16:09 web.1    |   app/helpers/entries_helper.rb:21:in 'EntriesHelper#entries_by_date'
12:16:09 web.1    |   app/views/transactions/index.html.erb:83
12:16:09 web.1    |   app/controllers/application_controller.rb:17:in 'ApplicationController#n_plus_one_detection'
12:16:09 web.1    |   app/controllers/concerns/localize.rb:17:in 'Localize#switch_timezone'
12:16:09 web.1    |   app/controllers/concerns/localize.rb:12:in 'Localize#switch_locale'
12:16:09 web.1    | 
12:16:09 web.1    | N+1 queries detected:
12:16:09 web.1    |   SELECT "accounts".* FROM "accounts" WHERE "accounts"."id" = $1 LIMIT $2
12:16:09 web.1    |   SELECT "accounts".* FROM "accounts" WHERE "accounts"."id" = $1 LIMIT $2
12:16:09 web.1    | Call stack:
12:16:09 web.1    |   app/models/transfer.rb:57:in 'Transfer#to_account'
12:16:09 web.1    |   app/models/transfer.rb:100:in 'Transfer#categorizable?'
12:16:09 web.1    |   app/views/transactions/_transaction_category.html.erb:4
12:16:09 web.1    |   app/views/transactions/_transaction.html.erb:96
12:16:10 web.1    |   app/views/transactions/_transaction.html.erb:6
12:16:10 web.1    |   app/views/transactions/_transaction.html.erb:5
12:16:10 web.1    |   app/views/entries/_entry.html.erb:3
12:16:10 web.1    |   app/views/transactions/index.html.erb:84
```

In total, I saw repeated SELECT calls for transactions, entries, and accounts, all triggered through nested partials in the transactions/index view.

[Skylight](https://oss.skylight.io/app/applications/XDpPIXEX52oi/recent/6h/endpoints/TransactionsController%23index?responseType=html) also highlighted this endpoint as a hot spot, and Prosopite helped confirm the source as N+1 queries:  
<img width="1266" height="766" alt="image" src="https://github.com/user-attachments/assets/dc578576-b691-42d7-9428-4577e6a549f0" />